Change: Extract a generic HTTP scanner library
