+++ http_scanner/libgvm_http_scanner.pc.in
@@ -1,0 +1,11 @@
+prefix=@CMAKE_INSTALL_PREFIX@
+exec_prefix=@EXEC_PREFIX@
+libdir=@LIBDIR@
+includedir=@INCLUDEDIR@
+
+Name: gvmlibs-http_scanner
+Description: Greenbone Vulnerability Management Library HTTP Scanner
+Version: @LIBGVMCONFIG_VERSION@
+Requires.private: glib-2.0 >= 2.42.0
+Cflags: -I${includedir} -I${includedir}/gvm
+Libs: -L${libdir} -lgvm_http_scanner
