+++ openvasd/openvasd_tests.c
@@ -20,225 +20,119 @@ AfterEach (openvasd)
 {
 }
 
-/* openvasd_delete_scan */
-
-Ensure (openvasd, openvasd_delete_scan_works_with_missing_id)
+Ensure (openvasd, openvasd_add_credential_to_scan_json)
 {
-  openvasd_resp_t resp;
-  openvasd_connector_t conn;
+  openvasd_credential_t *credential;
+  cJSON *credentials = cJSON_CreateArray ();
 
-  conn = openvasd_connector_new ();
-  resp = openvasd_delete_scan (conn);
-  assert_that (resp, is_not_null);
-  assert_that (resp->code, is_equal_to (RESP_CODE_ERR));
-  assert_that (resp->body,
-               is_equal_to_string ("{\"error\": \"Missing scan ID\"}"));
-  openvasd_response_cleanup (resp);
+  credential = openvasd_credential_new ("up", "generic", "0");
+
+  openvasd_credential_set_auth_data (credential, "username", "admin");
+  openvasd_credential_set_auth_data (credential, "password", "admin");
+
+  add_credential_to_scan_json (credential, credentials);
+
+  cJSON *credential_obj = cJSON_GetArrayItem (credentials, 0);
+  assert_that (cJSON_IsObject (credential_obj), is_true);
+
+  const char *service =
+    cJSON_GetStringValue (cJSON_GetObjectItem (credential_obj, "service"));
+  assert_that (service, is_equal_to_string ("generic"));
+
+  int port =
+    cJSON_GetNumberValue (cJSON_GetObjectItem (credential_obj, "port"));
+  assert_that (port, is_equal_to (0));
+
+  cJSON *auth_data = cJSON_GetObjectItem (credential_obj, "up");
+  const char *username =
+    cJSON_GetStringValue (cJSON_GetObjectItem (auth_data, "username"));
+  const char *password =
+    cJSON_GetStringValue (cJSON_GetObjectItem (auth_data, "password"));
+
+  assert_that (cJSON_IsObject (auth_data), is_true);
+  assert_that (username, is_equal_to_string ("admin"));
+  assert_that (password, is_equal_to_string ("admin"));
+
+  openvasd_credential_free (credential);
+  cJSON_Delete (credentials);
 }
 
-/* openvasd_stop_scan */
-
-Ensure (openvasd, openvasd_stop_scan_works_with_missing_id)
+Ensure (openvasd, openvasd_add_port_to_scan_json)
 {
-  openvasd_resp_t resp;
-  openvasd_connector_t conn;
+  range_t *ports_range;
+  cJSON *ports_range_array = cJSON_CreateArray ();
 
-  conn = openvasd_connector_new ();
-  resp = openvasd_stop_scan (conn);
-  assert_that (resp, is_not_null);
-  assert_that (resp->code, is_equal_to (RESP_CODE_ERR));
-  assert_that (resp->body,
-               is_equal_to_string ("{\"error\": \"Missing scan ID\"}"));
-  openvasd_response_cleanup (resp);
+  ports_range = g_malloc0 (sizeof (range_t));
+
+  ports_range->type = PORT_PROTOCOL_TCP;
+  ports_range->start = 22;
+  ports_range->end = 22;
+
+  add_port_to_scan_json (ports_range, ports_range_array);
+
+  cJSON *ports_obj = cJSON_GetArrayItem (ports_range_array, 0);
+
+  const char *protocol =
+    cJSON_GetStringValue (cJSON_GetObjectItem (ports_obj, "protocol"));
+
+  assert_that (cJSON_IsObject (ports_obj), is_true);
+  assert_that (protocol, is_equal_to_string ("tcp"));
+
+  cJSON *range_array = cJSON_GetObjectItem (ports_obj, "range");
+
+  assert_that (cJSON_IsArray (range_array), is_true);
+
+  cJSON *range_obj = cJSON_GetArrayItem (range_array, 0);
+
+  assert_that (cJSON_IsObject (range_obj), is_true);
+
+  int start = cJSON_GetNumberValue (cJSON_GetObjectItem (range_obj, "start"));
+  int end = cJSON_GetNumberValue (cJSON_GetObjectItem (range_obj, "end"));
+
+  assert_that (start, is_equal_to (22));
+  assert_that (end, is_equal_to (22));
+
+  g_free (ports_range);
+  cJSON_Delete (ports_range_array);
 }
 
-/* parse_results */
-
-Ensure (openvasd, parse_results_handles_details)
+Ensure (openvasd, openvasd_add_vts_to_scan_json)
 {
-  const gchar *str;
-  GSList *results;
-  openvasd_result_t result;
+  openvasd_vt_single_t *vt;
+  cJSON *vts_array = cJSON_CreateArray ();
 
-  results = NULL;
+  vt = openvasd_vt_single_new ("1.3.6.1.4.1.25623.1.0.877440");
 
-  str =
-    "[ {"
-    "  \"id\": 16,"
-    "  \"type\": \"host_detail\","
-    "  \"ip_address\": \"192.168.0.101\","
-    "  \"hostname\": \"g\","
-    "  \"oid\": \"1.3.6.1.4.1.25623.1.0.103997\","
-    "  \"message\": "
-    "\"<host><detail><name>MAC</name><value>94:E6:F7:67:4B:C0</"
-    "value><source><type>nvt</type><name>1.3.6.1.4.1.25623.1.0.103585</"
-    "name><description>Nmap MAC Scan</description></source></detail></host>\","
-    "  \"detail\": {"
-    "    \"name\": \"MAC\","
-    "    \"value\": \"00:1A:2B:3C:4D:5E\","
-    "    \"source\": {"
-    "      \"type\": \"nvt\","
-    "      \"name\": \"1.3.6.1.4.1.25623.1.0.103585\","
-    "      \"description\": \"Nmap MAC Scan\""
-    "    }"
-    "  }"
-    "} ]";
+  openvasd_vt_single_add_value (vt, "0", "bar");
 
-  parse_results (str, &results);
+  add_vts_to_scan_json (vt, vts_array);
 
-  assert_that (g_slist_length (results), is_equal_to (1));
+  cJSON *vts_obj = cJSON_GetArrayItem (vts_array, 0);
 
-  result = results->data;
-  assert_that (result->detail_name, is_equal_to_string ("MAC"));
-  assert_that (result->detail_value, is_equal_to_string ("00:1A:2B:3C:4D:5E"));
-  assert_that (result->detail_source_type, is_equal_to_string ("nvt"));
-  assert_that (result->detail_source_name,
-               is_equal_to_string ("1.3.6.1.4.1.25623.1.0.103585"));
-  assert_that (result->detail_source_description,
-               is_equal_to_string ("Nmap MAC Scan"));
+  assert_that (cJSON_IsObject (vts_obj), is_true);
 
-  if (g_slist_length (results))
-    g_slist_free_full (results, (GDestroyNotify) openvasd_result_free);
-}
+  const char *oid = cJSON_GetStringValue (cJSON_GetObjectItem (vts_obj, "oid"));
 
-/* parse_status */
+  assert_that (oid, is_equal_to_string ("1.3.6.1.4.1.25623.1.0.877440"));
 
-Ensure (openvasd, parse_status_start_end_time)
-{
-  const gchar *str;
-  openvasd_scan_status_t openvasd_scan_status = NULL;
+  cJSON *params_array = cJSON_GetObjectItem (vts_obj, "parameters");
 
-  openvasd_scan_status = g_malloc0 (sizeof (struct openvasd_scan_status));
-  str = "{"
-        "  \"start_time\":1737642308,"
-        "  \"end_time\":1737642389,"
-        "  \"status\":\"succeeded\","
-        "  \"host_info\":{"
-        "    \"all\":1,"
-        "    \"excluded\":0,"
-        "    \"dead\":0,"
-        "    \"alive\":1,"
-        "    \"queued\":0,"
-        "    \"finished\":1,"
-        "    \"scanning\":{},"
-        "    \"remaining_vts_per_host\":{}"
-        "  }"
-        "}";
+  assert_that (cJSON_IsArray (params_array), is_true);
 
-  parse_status (str, openvasd_scan_status);
+  cJSON *param_obj = cJSON_GetArrayItem (params_array, 0);
 
-  assert_that (openvasd_scan_status->status, is_equal_to (4));
-  assert_that_double (openvasd_scan_status->start_time,
-                      is_equal_to_double (1737642308));
-  assert_that_double (openvasd_scan_status->end_time,
-                      is_equal_to_double (1737642389));
+  assert_that (cJSON_IsObject (param_obj), is_true);
 
-  g_free (openvasd_scan_status);
-}
+  int id = cJSON_GetNumberValue (cJSON_GetObjectItem (param_obj, "id"));
 
-Ensure (openvasd, openvasd_connector_builder_all_valid_fields)
-{
-  openvasd_connector_t conn = openvasd_connector_new ();
+  const char *value =
+    cJSON_GetStringValue (cJSON_GetObjectItem (param_obj, "value"));
 
-  const char *ca_cert = "/path/to/ca.pem";
-  const char *cert = "/path/to/cert.pem";
-  const char *key = "/path/to/key.pem";
-  const char *apikey = "apikey-value";
-  const char *protocol = "https";
-  const char *host = "localhost";
-  const char *scan_id = "scan-uuid-123";
-  int port = 9390;
+  assert_that (value, is_equal_to_string ("bar"));
+  assert_that (id, is_equal_to (0));
 
-  assert_that (openvasd_connector_builder (conn, OPENVASD_CA_CERT, ca_cert),
-               is_equal_to (OPENVASD_OK));
-  assert_that (conn->ca_cert, is_equal_to_string (ca_cert));
-
-  assert_that (openvasd_connector_builder (conn, OPENVASD_CERT, cert),
-               is_equal_to (OPENVASD_OK));
-  assert_that (conn->cert, is_equal_to_string (cert));
-
-  assert_that (openvasd_connector_builder (conn, OPENVASD_KEY, key),
-               is_equal_to (OPENVASD_OK));
-  assert_that (conn->key, is_equal_to_string (key));
-
-  assert_that (openvasd_connector_builder (conn, OPENVASD_API_KEY, apikey),
-               is_equal_to (OPENVASD_OK));
-  assert_that (conn->apikey, is_equal_to_string (apikey));
-
-  assert_that (openvasd_connector_builder (conn, OPENVASD_PROTOCOL, protocol),
-               is_equal_to (OPENVASD_OK));
-  assert_that (conn->protocol, is_equal_to_string (protocol));
-
-  assert_that (openvasd_connector_builder (conn, OPENVASD_HOST, host),
-               is_equal_to (OPENVASD_OK));
-  assert_that (conn->host, is_equal_to_string (host));
-
-  assert_that (openvasd_connector_builder (conn, OPENVASD_SCAN_ID, scan_id),
-               is_equal_to (OPENVASD_OK));
-  assert_that (conn->scan_id, is_equal_to_string (scan_id));
-
-  assert_that (openvasd_connector_builder (conn, OPENVASD_PORT, &port),
-               is_equal_to (OPENVASD_OK));
-  assert_that (conn->port, is_equal_to (port));
-
-  g_free (conn->ca_cert);
-  g_free (conn->cert);
-  g_free (conn->key);
-  g_free (conn->apikey);
-  g_free (conn->protocol);
-  g_free (conn->host);
-  g_free (conn->scan_id);
-  g_free (conn);
-}
-
-Ensure (openvasd, openvasd_connector_builder_valid_protocol_http)
-{
-  openvasd_connector_t conn = openvasd_connector_new ();
-  openvasd_error_t result =
-    openvasd_connector_builder (conn, OPENVASD_PROTOCOL, "http");
-
-  assert_that (result, is_equal_to (OPENVASD_OK));
-  assert_that (conn->protocol, is_equal_to_string ("http"));
-
-  g_free (conn->protocol);
-  g_free (conn);
-}
-
-Ensure (openvasd, openvasd_connector_builder_invalid_protocol)
-{
-  openvasd_connector_t conn = openvasd_connector_new ();
-  openvasd_error_t result =
-    openvasd_connector_builder (conn, OPENVASD_PROTOCOL, "ftp");
-
-  assert_that (result, is_equal_to (OPENVASD_INVALID_VALUE));
-  assert_that (conn->protocol, is_null);
-
-  g_free (conn);
-}
-
-Ensure (openvasd, openvasd_connector_free)
-{
-  openvasd_connector_t conn = openvasd_connector_new ();
-
-  conn->ca_cert = g_strdup ("ca.pem");
-  conn->cert = g_strdup ("cert.pem");
-  conn->key = g_strdup ("key.pem");
-  conn->apikey = g_strdup ("api-key");
-  conn->protocol = g_strdup ("https");
-  conn->host = g_strdup ("localhost");
-  conn->scan_id = g_strdup ("scan-uuid");
-
-  openvasd_error_t result = openvasd_connector_free (conn);
-
-  assert_that (result, is_equal_to (OPENVASD_OK));
-}
-
-Ensure (openvasd, openvasd_connector_free_null_connector)
-{
-  openvasd_connector_t null_conn = NULL;
-  openvasd_error_t result = openvasd_connector_free (null_conn);
-
-  assert_that (result, is_equal_to (OPENVASD_OK));
+  openvasd_vt_single_free (vt);
+  cJSON_Delete (vts_array);
 }
 
 /* Test suite. */
@@ -250,23 +144,9 @@ main (int argc, char **argv)
 
   suite = create_test_suite ();
 
-  add_test_with_context (suite, openvasd, parse_results_handles_details);
-  add_test_with_context (suite, openvasd, parse_status_start_end_time);
-
-  add_test_with_context (suite, openvasd,
-                         openvasd_connector_builder_all_valid_fields);
-  add_test_with_context (suite, openvasd,
-                         openvasd_connector_builder_valid_protocol_http);
-  add_test_with_context (suite, openvasd,
-                         openvasd_connector_builder_invalid_protocol);
-  add_test_with_context (suite, openvasd, openvasd_connector_free);
-  add_test_with_context (suite, openvasd,
-                         openvasd_connector_builder_invalid_protocol);
-
-  add_test_with_context (suite, openvasd,
-                         openvasd_stop_scan_works_with_missing_id);
-  add_test_with_context (suite, openvasd,
-                         openvasd_delete_scan_works_with_missing_id);
+  add_test_with_context (suite, openvasd, openvasd_add_credential_to_scan_json);
+  add_test_with_context (suite, openvasd, openvasd_add_port_to_scan_json);
+  add_test_with_context (suite, openvasd, openvasd_add_vts_to_scan_json);
 
   if (argc > 1)
     ret = run_single_test (suite, argv[1], create_text_reporter ());
